python/02_financial_analyst/requirements.txt
openai
requests
python-dotenv
tabulate